requirements.txt
requests
googlesearch-python